Move flipr coordinator to its own file
