+++ homeassistant/components/flipr/__init__.py
@@ -1,27 +1,16 @@
 """The Flipr integration."""
-from datetime import timedelta
-import logging
-
-from flipr_api import FliprAPIRestClient
-from flipr_api.exceptions import FliprError
-
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_EMAIL, CONF_PASSWORD, Platform
+from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.update_coordinator import (
     CoordinatorEntity,
     DataUpdateCoordinator,
-    UpdateFailed,
 )
 
 from .const import ATTRIBUTION, CONF_FLIPR_ID, DOMAIN, MANUFACTURER
-
-_LOGGER = logging.getLogger(__name__)
-
-SCAN_INTERVAL = timedelta(minutes=60)
-
+from .coordinator import FliprDataUpdateCoordinator
 
 PLATFORMS = [Platform.BINARY_SENSOR, Platform.SENSOR]
 
@@ -47,38 +36,6 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class FliprDataUpdateCoordinator(DataUpdateCoordinator):
-    """Class to hold Flipr data retrieval."""
-
-    def __init__(self, hass, entry):
-        """Initialize."""
-        username = entry.data[CONF_EMAIL]
-        password = entry.data[CONF_PASSWORD]
-        self.flipr_id = entry.data[CONF_FLIPR_ID]
-
-        # Establishes the connection.
-        self.client = FliprAPIRestClient(username, password)
-        self.entry = entry
-
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=f"Flipr data measure for {self.flipr_id}",
-            update_interval=SCAN_INTERVAL,
-        )
-
-    async def _async_update_data(self):
-        """Fetch data from API endpoint."""
-        try:
-            data = await self.hass.async_add_executor_job(
-                self.client.get_pool_measure_latest, self.flipr_id
-            )
-        except FliprError as error:
-            raise UpdateFailed(error) from error
-
-        return data
 
 
 class FliprEntity(CoordinatorEntity):
