+++ tests/components/flipr/test_init.py
@@ -21,7 +21,7 @@ async def test_unload_entry(hass: HomeAssistant) -> None:
         unique_id="123456",
     )
     entry.add_to_hass(hass)
-    with patch("homeassistant.components.flipr.FliprAPIRestClient"):
+    with patch("homeassistant.components.flipr.coordinator.FliprAPIRestClient"):
         await hass.config_entries.async_setup(entry.entry_id)
         await hass.async_block_till_done()
         await hass.config_entries.async_unload(entry.entry_id)
